added new template, modified others
Created Task-list.Html for logged in users

index.php now serves a different template depending on if the user is already logged in. detection for such login will be at a later stage.

also fixed variable names to all follow the same rule of having a lowercase first letter and an uppercase second letter if need be.